Changed structure of training and trained some agents. (#85)
* New architecture versions of agents A2C and PPO trained a little.

* Removed A3C since the only one that will ever be trained and evaluated is A2C.

* Added paper on increasing dimensionality.

* Improved comments of quantum circuit enviornment.

* Removed normalization in favor of relative reductions.

* Removed normalization reward from training.

* Make ai pass selector print runtime params on --help.

* Refactored ai pass selector file.

* Trained 7 AI models.
