+++ AI/include/NeuralNetworks/Agents/A2C/base_a2c_agent.hpp
@@ -1,5 +1,5 @@
-#ifndef BASE_A3C_AGENT_HPP
-#define BASE_A3C_AGENT_HPP
+#ifndef BASE_A2C_AGENT_HPP
+#define BASE_A2C_AGENT_HPP
 
 // Torch includes
 #include "torch/torch.h"
@@ -7,35 +7,27 @@
 // Neural-Networks includes
 #include "NeuralNetworks/Agents/base_actor_critic.hpp"
 
-// Standard library includes
-#include <memory>
-
 namespace fs = std::filesystem;
 
 namespace ai_pass_selector {
 enum class OptimizerType : int;
 
-class BaseA3CAgent : public BaseActorCritic {
-protected:
-  /// A3C specific attributes
-  bool gradients_zero = true;
-  bool is_boss = true;
-
+class BaseA2CAgent : public BaseActorCritic {
 public:
   /// Constructors
-  explicit BaseA3CAgent(unsigned int max_qubits, bool is_boss = true);
+  explicit BaseA2CAgent(unsigned int max_qubits);
 
   /// Destructor
-  ~BaseA3CAgent() override = default;
+  ~BaseA2CAgent() override = default;
 
   /// Copy and move constructors and assignment operators
-  BaseA3CAgent(const BaseA3CAgent &other) noexcept = delete;
+  BaseA2CAgent(const BaseA2CAgent &other) noexcept = delete;
 
-  BaseA3CAgent(BaseA3CAgent &&other) noexcept = default;
+  BaseA2CAgent(BaseA2CAgent &&other) noexcept = default;
 
-  BaseA3CAgent &operator=(const BaseA3CAgent &other) noexcept = delete;
+  BaseA2CAgent &operator=(const BaseA2CAgent &other) noexcept = delete;
 
-  BaseA3CAgent &operator=(BaseA3CAgent &&other) noexcept = delete;
+  BaseA2CAgent &operator=(BaseA2CAgent &&other) noexcept = delete;
 
   /**
    * @param actor
@@ -46,7 +38,7 @@ public:
                   const torch::nn::Sequential &critic);
 
   //////////////////////////////////////////////////////////////////////////////
-  /// A2C/A3C standard methods
+  /// A2C/A2C standard methods
   /**
    * @param observation
    * @return
@@ -85,20 +77,8 @@ public:
    */
   void update_parameters(const torch::Tensor &actor_loss,
                          const torch::Tensor &critic_loss) const;
-
-  /**
-   * Necessary to create worker agents for asynchronous training.
-   * @return
-   */
-  virtual std::unique_ptr<BaseA3CAgent> clone() const = 0;
-  void zero_grad();
-  void load_weights(BaseA3CAgent &other);
-  void load_gradients(BaseA3CAgent &other);
-  void update_parameters_assuming_gradients_are_loaded();
   //////////////////////////////////////////////////////////////////////////////
-  /// Saving and Loading
-  void save_model() const override;
 };
 } // namespace ai_pass_selector
 
-#endif // BASE_A3C_AGENT_HPP
+#endif // BASE_A2C_AGENT_HPP
