--- AI/include/NeuralNetworks/Agents/A3C/a3c_agents.hpp
@@ -1,31 +1,0 @@
-#ifndef A3C_AGENTS_HPP
-#define A3C_AGENTS_HPP
-
-// Torch includes
-#include "NeuralNetworks/Agents/A3C/base_a3c_agent.hpp"
-
-#define DECLARE_A3C_AGENT(ClassName)                                           \
-  class ClassName final : public BaseA3CAgent {                                \
-  public:                                                                      \
-    explicit ClassName(unsigned int max_qubits, bool is_boss = true);          \
-    std::string agentName() const override;                                    \
-    std::unique_ptr<BaseA3CAgent> clone() const override;                      \
-  };
-
-namespace ai_pass_selector {
-/// A2C = Advantage Actor-Critic
-/// A3C = Asynchronous Advantage Actor-Critic
-/// TCN = Temporal Convolutional Network
-/// LSTM = Long Short-Term Memory
-/// HYBRID = Hybrid of TCN and LSTM
-// a3c-mq28-tcn Nr trainable parameters:
-// a3c-mq130-tcn Nr trainable parameters:
-DECLARE_A3C_AGENT(A3C_TCN)
-// a3c-mq28-lstm Nr trainable parameters:
-// a3c-mq130-lstm Nr trainable parameters:
-DECLARE_A3C_AGENT(A3C_LSTM)
-// a3c-mq28-hybrid Nr trainable parameters:
-// a3c-mq130-hybrid Nr trainable parameters:
-DECLARE_A3C_AGENT(A3C_HYBRID)
-} // namespace ai_pass_selector
-#endif // A3C_AGENTS_HPP
