+++ AI/include/NeuralNetworks/Agents/ACER/base_acer_agent.hpp
@@ -106,7 +106,6 @@ public:
   );
 
   /**
-   *
    * @param actor_gradients
    * @param critic_loss
    */
