+++ AI/include/NeuralNetworks/Agents/agent_utils.hpp
@@ -18,7 +18,7 @@ using namespace mqss::opt;
 
 namespace ai_pass_selector {
 enum class AgentClass : int {
-  A3C = 1,
+  A2C = 1,
   PPO = 2,
   SDSAC = 3,
   ACER = 4,
@@ -46,14 +46,14 @@ struct AgentAttributes {
 };
 
 const std::unordered_map<std::string, AgentClass> AGENT_NAME_TO_CLASS = {
-    {"a3c", AgentClass::A3C},
+    {"a2c", AgentClass::A2C},
     {"ppo", AgentClass::PPO},
     {"acer", AgentClass::ACER},
     {"sdsac", AgentClass::SDSAC},
     {"crossq", AgentClass::CROSSQ}};
 
 const std::unordered_map<AgentClass, std::string, EnumClassHash>
-    AGENT_CLASS_TO_NAME = {{AgentClass::A3C, "a3c"},
+    AGENT_CLASS_TO_NAME = {{AgentClass::A2C, "a2c"},
                            {AgentClass::PPO, "ppo"},
                            {AgentClass::ACER, "acer"},
                            {AgentClass::SDSAC, "sdsac"},
